Reading files and bug fixed
Added reading and writting to binary files

+++ Simulator3D/Simulator3D/IO/WriteSBF.h
@@ -18,7 +18,7 @@ public:
 
 private:
 
-	static constexpr unsigned int size_bulk = 100;
+	static constexpr unsigned long size_bulk = 100;
 
 	const unsigned long n;
 	const unsigned long rest;
